# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: captions
    description: Caption generation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /captions/generate:
    post:
      operationId: generateCaptions
      tags: [captions]
      summary: Generate social media captions using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCaptionsBody"
      responses:
        "200":
          description: Generated captions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateCaptionsResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /captions/regenerate-one:
    post:
      operationId: regenerateOneCaption
      tags: [captions]
      summary: Regenerate a single caption
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegenerateOneCaptionBody"
      responses:
        "200":
          description: Regenerated caption
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptionItem"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /captions/hashtags:
    post:
      operationId: generateHashtags
      tags: [captions]
      summary: Generate trending hashtags for a niche and topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateHashtagsBody"
      responses:
        "200":
          description: Generated hashtags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateHashtagsResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    GenerateCaptionsBody:
      type: object
      properties:
        niche:
          type: string
          enum:
            - Real Estate
            - Fitness Coach
            - Restaurant
            - Boutique/Shop
            - General Business
        postDescription:
          type: string
        tone:
          type: string
          enum:
            - Professional
            - Casual
            - Funny
            - Inspirational
        platform:
          type: string
          enum:
            - Instagram
            - Facebook
            - LinkedIn
            - TikTok
            - Twitter/X
        postType:
          type: string
          enum:
            - Product Showcase
            - Sale/Promo
            - Behind the Scenes
            - Tips & Education
            - Announcement
            - Customer Story
        captionLength:
          type: string
          enum:
            - Short
            - Medium
            - Long
        includeEmojis:
          type: boolean
        ctaType:
          type: string
          enum:
            - None
            - Shop Now
            - Link in Bio
            - DM Us
            - Comment Below
            - Tag a Friend
            - Save This Post
      required:
        - niche
        - postDescription
        - tone
    CaptionItem:
      type: object
      properties:
        caption:
          type: string
        hashtags:
          type: string
      required:
        - caption
        - hashtags
    GenerateCaptionsResponse:
      type: object
      properties:
        captions:
          type: array
          items:
            $ref: "#/components/schemas/CaptionItem"
      required:
        - captions
    RegenerateOneCaptionBody:
      type: object
      properties:
        niche:
          type: string
        postDescription:
          type: string
        tone:
          type: string
        platform:
          type: string
        postType:
          type: string
        captionLength:
          type: string
        includeEmojis:
          type: boolean
        ctaType:
          type: string
        existingCaptions:
          type: array
          items:
            type: string
          description: Captions already shown so the new one is different
      required:
        - niche
        - postDescription
        - tone
    GenerateHashtagsBody:
      type: object
      properties:
        niche:
          type: string
        topic:
          type: string
        platform:
          type: string
      required:
        - niche
        - topic
    GenerateHashtagsResponse:
      type: object
      properties:
        hashtags:
          type: array
          items:
            type: string
        grouped:
          type: object
          properties:
            niche:
              type: array
              items:
                type: string
            trending:
              type: array
              items:
                type: string
            broad:
              type: array
              items:
                type: string
          required:
            - niche
            - trending
            - broad
      required:
        - hashtags
        - grouped
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error