# lib/api-zod/src/generated/types/generateHashtagsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateHashtagsResponseGrouped } from "./generateHashtagsResponseGrouped";

export interface GenerateHashtagsResponse {
  hashtags: string[];
  grouped: GenerateHashtagsResponseGrouped;
}