# lib/api-zod/src/generated/types/regenerateOneCaptionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegenerateOneCaptionBody {
  niche: string;
  postDescription: string;
  tone: string;
  platform?: string;
  postType?: string;
  captionLength?: string;
  includeEmojis?: boolean;
  ctaType?: string;
  /** Captions already shown so the new one is different */
  existingCaptions?: string[];
}